requirements.txt
langchain
langchain-google-genai
langgraph
arxiv
pypdf
requests
reportlab
streamlit
python-dotenv
